// PlanItKorea_w/src/types/type.ts
export type Announcement = {
  id: number;
  boardType: BoardType;
  boardTitle: string;
  boardContent: string;
  author: string;
  uploadDate: Date;
};

export type BoardType = "공지사항" | "자주묻는질문";